requirements.txt
chardet==2.3.0
click==7.1.2
cryptography==1.7.1
goprocam==4.1.0
idna==2.2
keyring==10.1
keyrings.alt==1.3
pbkdf2==1.3
pyasn1==0.1.9
pycrypto==2.6.1
pygobject==3.22.0
python-apt==1.1.0b5
pyxdg==0.25
requests==2.12.4
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
ssh-import-id==5.6
urllib3==1.19.1
wifi==0.3.8
